pyproject.toml
[project]
name = "lmsk-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "requests>=2.32.3",
    "beautifulsoup4==4.13.2",
    "python-dotenv>=1.0.1",
    "icalendar>=7.0.3",
]
